frontend(onboardingForm): rebase and removed onboardingFormWrapper
