+++ .vscode/settings.json
@@ -1,6 +1,6 @@
 {
   "eslint.useESLintClass": true,
   "editor.codeActionsOnSave": {
-    "source.fixAll.eslint": true
+    "source.fixAll.eslint": "explicit"
   }
 }
