+++ frontend/package.json
@@ -14,17 +14,17 @@
     "prepare": "cd .. && husky"
   },
   "dependencies": {
-    "little-state-machine": "^5.0.1",
     "i18next": "^24.2.1",
     "i18next-browser-languagedetector": "^8.0.2",
+    "little-state-machine": "^5.0.1",
     "next": "15.1.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-hook-form": "^7.54.2",
     "react-i18next": "^15.4.0",
+    "react-router-dom": "^7.1.3",
     "react-select": "^5.10.0",
     "react-select-country-list": "^2.2.3"
-    "react-router-dom": "^7.1.3"
-    "react-hook-form": "^7.54.2",
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
