+++ frontend/src/components/onboardingForm/index.ts
@@ -3,4 +3,3 @@ export { Step1 } from "./Step1";
 export { Step2 } from "./Step2";
 export { Step3A } from "./Step3A";
 export { Step3B } from "./Step3B";
-export { OnboardingFormWrapper } from "./onboardingFormWrapper";
